add relationship between school and  student

+++ src/models/TeacherList.kt
@@ -4,5 +4,5 @@ data class TeacherList(
     val id: String ,
     var name: String,
     var teachingSpecialization: String,
-    var phone: Int,
+    var phone: Long,
 )
